Replace must.NotFail(types.NewDocument()) => types.NewEmptyDocument()
Added a new helper function, `types.NewDocument`.
This function is identical to `types.MakeDocument(0)`, but it
doesn't require the user to specify the capacity.

The main benefit is making the simple operation of creating a
new (empty) document easier to read (not `must` package involved).

Some use cases can be further improved by using an appropriate
size hints (see TODO comments), but I tried to avoid the unrelated
changes in this PR (plus it doesn't matter if these use sites are
not on the hot path).